requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
python-dateutil
jinja2
python-multipart
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose[cryptography]
passlib
authlib
httpx
itsdangerous